store original and intermediate traces
